+++ src/scvi/module/_vaec.py
@@ -197,7 +197,11 @@ class VAEC(BaseModuleClass):
         h = self.decoder(*decoder_input)
         px_scale = self.px_decoder(h)
         px_rate = library * px_scale
-        return {MODULE_KEYS.PX_KEY: NegativeBinomial(px_rate, logits=self.px_r)}
+        return {
+            MODULE_KEYS.PX_KEY: NegativeBinomial(
+                px_rate, logits=self.px_r, on_mps=(self.device.type == "mps")
+            )
+        }
 
     def loss(
         self,
@@ -262,7 +266,7 @@ class VAEC(BaseModuleClass):
         px_r = generative_outputs["px_r"]
         px_rate = generative_outputs["px_rate"]
 
-        dist = NegativeBinomial(px_rate, logits=px_r)
+        dist = NegativeBinomial(px_rate, logits=px_r, on_mps=(self.device.type == "mps"))
         if n_samples > 1:
             exprs = dist.sample().permute([1, 2, 0])  # Shape : (n_cells_batch, n_genes, n_samples)
         else: